requirements.txt
Flask
flask_cors
flask_sqlalchemy
opencv-python
numpy
Pillow
ultralytics
gunicorn